Moved auth initializers to their own controller
